requirements.txt
boto3==1.11.10
botocore==1.14.10
Django==2.2.9
django-cors-headers==3.2.1
django-filter==2.2.0
django-s3-storage==0.13.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
Markdown==3.1.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.2
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
views==0.3
